Fix for building ref-ui image
Since the Dockerfile does not specify the version of the nginx
parent, we are at their mercy. Seems after some updates to nginx
and debian, gnupg is no longer installed with apt-get.

I added gnupg install to the Dockerfile, which solved the build issue
on systems that do not have old nginx versions present. However we should
most likely settle on an nginx version for stability.